+++ lavacli/__init__.py
@@ -1,3 +1,3 @@
 """LavaCLI - A Beautiful Terminal Lava Lamp."""
 
-__version__ = '1.5.0'
+__version__ = '1.6.0'